Move ChannelContentConfiguration into `types/`
