+++ packages/shared/src/api/index.ts
@@ -1,4 +1,3 @@
-export * from './channelContentConfig';
 export * from './channelsApi';
 export * from './chatApi';
 export * from './contactsApi';
