+++ packages/shared/src/db/schema.ts
@@ -8,7 +8,7 @@ import {
   uniqueIndex,
 } from 'drizzle-orm/sqlite-core';
 
-import { ChannelContentConfiguration } from '../api';
+import { ChannelContentConfiguration } from '../types/ChannelContentConfiguration';
 import { ExtendedEventType, NotificationLevel, Rank } from '../urbit';
 
 const boolean = (name: string) => {
