+++ packages/shared/src/types/ChannelContentConfiguration.ts
