+++ packages/ui/src/components/Channel/PostView.tsx
@@ -1,5 +1,4 @@
-import { JSONValue } from '@tloncorp/shared';
-import { ChannelContentConfiguration } from '@tloncorp/shared/api';
+import { ChannelContentConfiguration, JSONValue } from '@tloncorp/shared';
 import { useMemo } from 'react';
 
 import { useChannelContext } from '../../contexts';
