+++ packages/ui/src/components/Channel/index.tsx
@@ -4,11 +4,9 @@ import {
   useGroupPreview,
   usePostReference as usePostReferenceHook,
   usePostWithRelations,
-} from '@tloncorp/shared';
-import {
   ChannelContentConfiguration,
   DraftInputId,
-} from '@tloncorp/shared/api';
+} from '@tloncorp/shared';
 import * as db from '@tloncorp/shared/db';
 import { JSONContent, Story } from '@tloncorp/shared/urbit';
 import { ImagePickerAsset } from 'expo-image-picker';
